utilities/requirements.txt
gspread 
oauth2client
reportlab 
python-barcode 
pillow